move MapboxDistanceFormatter to a public package
